Move `Join` extension to `IEnumerable` class
Moved the extensions and tests from the `string` class to the `IEnumerable` class. Updated the `ReadMe.md` file.
