+++ src/Useful.Extensions/StringExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text.RegularExpressions;
@@ -408,33 +407,5 @@ namespace Useful.Extensions
         }
 
         #endregion IsAllAlphaOrNumbers
-
-        #region String.Join
-
-        public static string Join(this IEnumerable<string> source, char separator)
-        {
-            var result = string.Join(separator, source);
-            return result;
-        }
-
-        public static string Join(this IEnumerable<string> source, string separator)
-        {
-            var result = string.Join(separator, source);
-            return result;
-        }
-
-        public static string Join<T>(this IEnumerable<T> source, char separator)
-        {
-            var result = string.Join(separator, source);
-            return result;
-        }
-
-        public static string Join<T>(this IEnumerable<T> source, string separator)
-        {
-            var result = string.Join(separator, source);
-            return result;
-        }
-        
-        #endregion String.Join
     }
 }
